refactor(nodebuilder/fraud): Move fraud-related constructors into separate module
