+++ nodebuilder/header/header.go
@@ -3,14 +3,10 @@ package header
 import (
 	"context"
 
-	"github.com/ipfs/go-datastore"
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/libp2p/go-libp2p-core/peerstore"
-	pubsub "github.com/libp2p/go-libp2p-pubsub"
-	"go.uber.org/fx"
 
-	"github.com/celestiaorg/celestia-node/fraud"
 	"github.com/celestiaorg/celestia-node/header"
 	"github.com/celestiaorg/celestia-node/header/p2p"
 	"github.com/celestiaorg/celestia-node/header/store"
@@ -50,72 +46,5 @@ func InitStore(ctx context.Context, cfg Config, net params.Network, s header.Sto
 		log.Errorf("initializing store failed: %s", err)
 	}
 
-	return nil
-}
-
-// FraudService constructs a fraud proof service with the syncer disabled.
-func FraudService(
-	lc fx.Lifecycle,
-	sub *pubsub.PubSub,
-	host host.Host,
-	hstore header.Store,
-	ds datastore.Batching,
-) (fraud.Service, error) {
-	return newFraudService(lc, sub, host, hstore, ds, false)
-}
-
-// FraudServiceWithSyncer constructs fraud proof service with enabled syncer.
-func FraudServiceWithSyncer(
-	lc fx.Lifecycle,
-	sub *pubsub.PubSub,
-	host host.Host,
-	hstore header.Store,
-	ds datastore.Batching,
-) (fraud.Service, error) {
-	return newFraudService(lc, sub, host, hstore, ds, true)
-}
-
-func newFraudService(
-	lc fx.Lifecycle,
-	sub *pubsub.PubSub,
-	host host.Host,
-	hstore header.Store,
-	ds datastore.Batching,
-	isEnabled bool) (fraud.Service, error) {
-	pservice := fraud.NewProofService(sub, host, hstore.GetByHeight, ds, isEnabled)
-	lc.Append(fx.Hook{
-		OnStart: pservice.Start,
-		OnStop:  pservice.Stop,
-	})
-	return pservice, nil
-}
-
-// FraudLifecycle controls the lifecycle of service depending on fraud proofs.
-// It starts the service only if no fraud-proof exists and stops the service automatically
-// if a proof arrives after the service was started.
-func FraudLifecycle(
-	startCtx, lifecycleCtx context.Context,
-	p fraud.ProofType,
-	fservice fraud.Service,
-	start, stop func(context.Context) error,
-) error {
-	proofs, err := fservice.Get(startCtx, p)
-	switch err {
-	default:
-		return err
-	case nil:
-		return &fraud.ErrFraudExists{Proof: proofs}
-	case datastore.ErrNotFound:
-	}
-	err = start(startCtx)
-	if err != nil {
-		return err
-	}
-	// handle incoming Fraud Proofs
-	go fraud.OnProof(lifecycleCtx, fservice, p, func(fraud.Proof) {
-		if err := stop(lifecycleCtx); err != nil {
-			log.Error(err)
-		}
-	})
 	return nil
 }
